file upload function + backend pdf generation
