+++ lego_django/lego/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
 
+from .models import ArtProject
+
 # Register your models here.
+
+admin.site.register(ArtProject)
