+++ lego_django/templates/lego/base.html
@@ -13,17 +13,17 @@
     <!-- Navbar -->
     <nav class="navbar navbar-expand-lg bg-dark navbar-dark">
     <div class="container">
-        <a href="#" class="navbar-brand">Lego Art Project</a>
+        <a href="" class="navbar-brand">Lego Art Project</a>
         <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navmenu">
             <span class="navbar-toggler-icon"></span>
         </button>
         <div class="collapse navbar-collapse" id="navmenu">
             <ul class="navbar-nav ms-auto">
                 <li class="nav-item">
-                    <a href="#" class="nav-link">Project</a>
+                    <a href="" class="nav-link">Project</a>
                 </li>
                 <li class="nav-item">
-                    <a href="about" class="nav-link">About</a>
+                    <a href="#about/" class="nav-link">About</a>
                 </li>
             </ul>
         </div>
